requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
python-dotenv
mlflow
astrapy